requirements/base.txt
Django==1.8.6
Fabric==1.10.1
gunicorn==19.3.0
httplib2==0.9.1
Pillow==3.0.0
psycopg2==2.6
requests==2.8.1
requests-oauthlib==0.5.0
django-background-tasks